add tests

+++ ml_project/enities/splitting_params.py
@@ -3,5 +3,5 @@ from dataclasses import dataclass
 
 @dataclass()
 class SplittingParams:
-    val_size: int
+    val_size: float
     random_state: int
